[chore] move lazy &&~ and lazy ||~ to App module
